# zobba/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zobba owned HTTP interface",
    "description": "Service health, current scoped identity, durable Task commands and exact operation decisions under standing Permissions. Provider acceptance and completed effects are distinct; no live connector, model, computer or audit execution is implied.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/auth/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "callback",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Exact one-use state from the current login attempt",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 43,
              "minLength": 43,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "One-use provider authorization code; required unless error is present",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096,
              "minLength": 1
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "Provider denial code; never reflected to clients",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Verified success rotates session; failure returns only fixed sign-in status to configured app. Clear only matching consumed binding."
          },
          "400": {
            "description": "Invalid response when no application origin is configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Whole-request deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "login_binding": []
          }
        ]
      }
    },
    "/auth/login": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "responses": {
          "303": {
            "description": "Redirect to trusted OIDC sign-in with one-use state and PKCE"
          },
          "429": {
            "description": "Outstanding sign-in capacity reached; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Recoverable sign-in failure; native browser navigation receives an HTML return link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Exactly the configured HTTPS application origin, supplied by the browser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Current session-bound CSRF token from GET /auth/session",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 43,
              "minLength": 43,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Server session deleted and cookie expired"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "server_session": []
          }
        ]
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "session",
        "parameters": [
          {
            "name": "X-Expected-Session",
            "in": "header",
            "description": "Optional session-bound read precondition from the in-memory session CSRF token; mismatch refuses without changing the cookie",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "maxLength": 128,
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Session changed; compose fresh reads without replacing the current cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "server_session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IdentityResponse": {
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "identity",
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityResponse"
          }
        }
      },
...
    },
    "securitySchemes": {
      "login_binding": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-zobba-login",
        "description": "One-use current login browser binding; Secure, HttpOnly, SameSite=Lax, Path=/"
      },
      "server_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-zobba-session",
        "description": "Opaque server session; Secure, HttpOnly, SameSite=Lax, Path=/; browser managed, never an OIDC provider token"
      }
    }
  }
}